fix: address code review findings
- Split lock.go into platform-specific files with build tags to fix
  Windows build (syscall.Flock is Unix-only)
- Remove dead code in clone_service.go (unreachable dry-run checks)
- Remove "table" from config list --format help text (not implemented)
- Use t.Setenv instead of os.Setenv in config tests
- Assert NoError in clone dry-run test instead of swallowing error
