+++ internal/manifest/lock_test.go
@@ -1,3 +1,5 @@
+//go:build unix
+
 package manifest
 
 import (
